+++ cpe_interface_struct.go
@@ -30,6 +30,9 @@ type UciLbsCfg struct {
 	ModelName      string   `json:"model_name"`
 	MacList        []string `json:"maclist"`
 	MacFilter      string   `json:"macfilter"`
+	MaxQuiet       string   `json:"max_quiet"`
+	MaxCacheQueue  string   `json:"max_cache_queue"`
+	EmptyWatcher   string   `json:"empty_watcher"`
 	Type           string   `json:".type,omitempty"`
 }
 type UciStatCfg struct {
@@ -72,6 +75,19 @@ type UciFirmwareCfg struct {
 	CurrentMD5   string `json:"current_md5,omitempty"`
 	AvailableMD5 string `json:"available_md5,omitempty"`
 }
+
+type UciWsnmpd struct {
+	Default UciWsnmpdDefautl `json:"default,omitempty"`
+}
+
+type UciWsnmpdDefautl struct {
+	Enabled         string   `json:"enabled"`
+	Community       string   `json:"community"`
+	Location        string   `json:"location"`
+	ListenInterface string   `json:"listen_interface"`
+	Interfaces      []string `json:"interfaces"`
+}
+
 type UciEthernetAcct struct {
 	Type       string   `json:".type,omitempty"`
 	Interfaces []string `json:"interface,omitempty"`
@@ -219,11 +235,13 @@ type UciWifiWlan struct {
 	WmmAcVOAcm   string `json:"wmm_ac_vo_acm"`
 
 	// speed limit
-
 	// limit_fromrf -- FROM client -- UPLOAD
 	// limit_torf -- TO client -- DOWNLOAD
 	LimitFromRf string `json:"limit_fromrf"`
 	LimitToRf   string `json:"limit_torf"`
+
+	// timeouts
+	MaxInactivity string `json:"max_inactivity"`
 }
 type UciWifiIface struct {
 	Type           string      `json:".type"`
@@ -240,6 +258,7 @@ type UciWifiIface struct {
 	SupportedRates interface{} `json:"supported_rates,omitempty"`
 	LegacyRates    string      `json:"legacy_rates,omitempty"`
 	LogLevel       string      `json:"log_level,omitempty"`
+	MaxInactivity  string      `json:"max_inactivity,omitempty"`
 }
 type UciRadius struct {
 	Type       string `json:".type"`
@@ -333,6 +352,7 @@ type UciConfig struct {
 	Wireless UciWireless `json:"wireless,omitempty"`
 	System   UciSystem   `json:"system,omitempty"`
 	Network  UciNetwork  `json:"network,omitempty"`
+	Wsnmpd   UciWsnmpd   `json:"wmsnmpd,omitempty"`
 }
 
 //------------- JSON conversion ----------------