+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## current
+
+- WNE-2047: Add: MaxInactivity timeout to WiFiConfig (8c01a7e)
+
 ## [v1.5.0] - 08-10-2020
  - WNE-1925: Add: add new fields in UciLbsCfg and LBSConfig (0d91ccf)
  - WNE-1976: Fix: maligned/errcheck errors with golangci-lint (2b50ae3)