+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [v1.5.0] - 08-10-2020
- - WNE-2060: Add: new field first connection in cpe model ()
+ - WNE-2060: Add: new field first connection in cpe model (dcf6654)
  - WNE-1976: Fix: maligned/errcheck errors with golangci-lint (2b50ae3)
  - WNE-1976: Delete: go.sum with append to gitignore (5d3ee37)
  - WNE-1918: Delete: func (69be4f8)