+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## current
 
-- WNE-2047: Add: MaxInactivity timeout to WiFiConfig (8c01a7e)
+ - WNE-2060: Add: new field first connection in cpe model (dcf6654)
+ - WNE-2047: Add: MaxInactivity timeout to WiFiConfig (8c01a7e)
 
 ## [v1.5.0] - 08-10-2020
+
  - WNE-1925: Add: add new fields in UciLbsCfg and LBSConfig (0d91ccf)
  - WNE-1976: Fix: maligned/errcheck errors with golangci-lint (2b50ae3)
  - WNE-1976: Delete: go.sum with append to gitignore (5d3ee37)