+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## current
+ - WNE-2059: Add: snmp module name (fa9b531)
  - WNE-2055: Add: wmsnmpd config in cpe and uci model (f73ae6c)
  - WNE-2060: Add: new field first connection in cpe model (dcf6654)
  - WNE-2047: Add: MaxInactivity timeout to WiFiConfig (8c01a7e)