+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## current
+## [v1.6.0]
+
+ - WNE-2361: Add: MapVendor changed from string to int (1811d0f)
  - WNE-2314: Add: RSSI threshold and Band steering (d9f880b)
  - WNE-2276: Add: 802.11v and 802.11k (1ffa1e2)
  - WNE-2161: Add: new module resampling (bd401f1)